Reorganize simulator logic into a subpackage
