+++ magma/simulator/__init__.py
@@ -1,0 +1,2 @@
+import magma.simulator.mdb
+from .python_simulator import PythonSimulator
