+++ magma/simulator/mdb.py
@@ -1,13 +1,13 @@
 from __future__ import print_function
 from .python_simulator import PythonSimulator
-from .transforms import get_uniq_circuits
-from .circuit import *
-from .scope import *
-from .array import ArrayType
-from .bit import BitType
-from .bitutils import seq2int
-from .ref import InstRef
-from .compatibility import builtins
+from ..transforms import get_uniq_circuits
+from ..circuit import *
+from ..scope import *
+from ..array import ArrayType
+from ..bit import BitType
+from ..bitutils import seq2int
+from ..ref import InstRef
+from ..compatibility import builtins
 from code import compile_command
 import re
 from sys import platform
