+++ tests/test_expressions.py
@@ -4,7 +4,7 @@ from magma.verilator.verilator import compile as compileverilator
 from magma.verilator.verilator import run_verilator_test
 from magma.verilator.function import testvectors
 import itertools
-from magma.python_simulator import PythonSimulator
+from magma.simulator import PythonSimulator
 from magma.scope import Scope
 
 def test_1_bit_logic():
