+++ tests/test_primitives.py
@@ -15,7 +15,7 @@ from magma.bitutils import int2seq
 from magma.verilator.verilator import compile as compileverilator
 from magma.verilator.verilator import run_verilator_test
 from magma.verilator.function import testvectors
-from magma.python_simulator import PythonSimulator
+from magma.simulator import PythonSimulator
 from magma.scope import Scope
 from magma.verilator.function import testvectors
