+++ tests/test_simulator/test_counter.py
@@ -1,5 +1,5 @@
 from .test_primitives import *
-from magma.python_simulator import PythonSimulator
+from magma.simulator import PythonSimulator
 from magma import *
 from magma.scope import *
