+++ tests/test_simulator/test_logic.py
@@ -1,5 +1,5 @@
 from .test_primitives import *
-from magma.python_simulator import PythonSimulator
+from magma.simulator import PythonSimulator
 from magma.scope import *
 
 def test():
